add move method for Snake

+++ core/src/com/nickmillward/snake/entities/Point.java
@@ -5,31 +5,31 @@ package com.nickmillward.snake.entities;
  */
 public class Point {
 
-    private int x, y;
+    private float x, y;
 
     public Point() {
         x = 0;
         y = 0;
     }
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 }
